fixed a minor bug

+++ Experiments_Engine/Policies/Epsilon_Greedy.py
@@ -25,16 +25,17 @@ class EpsilonGreedyPolicy(PolicyBase):
         if True in (np.array(q_value) == np.inf):
             raise ValueError("One of the Q-Values has a value of infinity.")
         if p < self.epsilon:
-            action = randint(self.numActions)
+            action = randint(self.numActions, dtype=np.uint8)
         else:
-            action = np.random.choice(np.argwhere(q_value == np.max(q_value)).flatten(), size=1)[0]
+            # choosing a random action from all the possible maximum action
+            action = np.uint8(np.random.choice(np.argwhere(q_value == np.max(q_value)).flatten(), size=1)[0])
         return action
 
     """" Returns the probability of a given action or of all the actions """
     def probability_of_action(self, q_values, action=0, all_actions=False):
         max_q = max(q_values)
         total_max_actions = sum(max_q == array(q_values))
-        action_probabilities = zeros(self.numActions)
+        action_probabilities = zeros(self.numActions, dtype=np.float64)
 
         for i in range(self.numActions):
             if q_values[i] == max_q:
